pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "litescript"
version = "0.0.1"
description = "Utility package for streamlined Python scripts with logging, common imports, and essential utilities"
readme = "README.rst"
authors = [
  {name = "Somesh Fengade", email = "someshfengde@gmail.com"}
]
maintainers = [
  {name = "Somesh Fengade", email = "someshfengde@gmail.com"}
]
classifiers = [

]
license = {text = "MIT license"}
dependencies = [
  "tqdm", 
  "colorlog",
  "pyyaml", 
  "rich", 
  "numpy", 
  "pandas",
  "plotly", 
  "matplotlib"
]

[project.optional-dependencies]
dev = [
    "coverage",  # testing
    "mypy",  # linting
    "pytest",  # testing
    "ruff"  # linting
]

[project.urls]

bugs = "https://github.com/someshfengde/litescript/issues"
changelog = "https://github.com/someshfengde/litescript/blob/master/changelog.md"
homepage = "https://github.com/someshfengde/litescript"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
"*" = ["*.*"]




# Mypy
# ----

[tool.mypy]
files = "."

# Use strict defaults
strict = true
warn_unreachable = true
warn_no_return = true

[[tool.mypy.overrides]]
# Don't require test functions to include types
module = "tests.*"
allow_untyped_defs = true
disable_error_code = "attr-defined"

